requirements.txt
# YOLOv5 requirements
# Usage: pip install -r requirements.txt

# Base ------------------------------------------------------------------------
numpy>=1.22.2
seaborn>=0.13.0
opencv-python>=4.1.1
Pillow>=10.0.1
psutil  # system resources
PyYAML>=5.3.1
requests>=2.23.0
scipy>=1.4.1
thop>=0.1.1  # FLOPs computation
torch>=1.8.0  # see https://pytorch.org/get-started/locally (recommended)
torchvision>=0.9.0
tqdm>=4.64.0
ultralytics>=8.0.147
# protobuf<=3.20.1

# Logging ---------------------------------------------------------------------
# tensorboard>=2.4.1

pandas>=1.1.4
tensorflow>=2.13.0
openpyxl>=3.1.2
joblib>=1.3.2
scikit-learn>=1.3.2
albumentations>=1.3.1
fsspec

tensorflow-addons

ipython  # interactive notebook
ipykernel
# python3-opencv
# opencv-python
# opencv-contrib-python
# opencv-python-headless==4.8.1.78
opencv-python-headless==4.5.3.56
# typing-extensions>=4.7.1
# tensorrt
# tensorflow
# tensorflow-gpu == 2.3.3



